ADAS-Stereo-Vision/requirements.txt
numpy
pillow
scikit-learn
tqdm
albumentations
jupyterlab
matplotlib
natsort
scikit-image>=0.16.1
tensorboardx
tensorboard
torchcontrib
tifffile
pygit2